fix datepicker
